Read board extensions
Still, these are trapped into the map loader and cannot be brought
out... yet. I guess the simpler way of doing this is just to make a
getter, reset before reading and be done. Actually it may just pay out
to have copies of both the board and this data on the driver, to allow
for speedy resets.